requirements.txt
Django==1.5.4
MySQL-python==1.2.4
cffi==0.6
django-grappelli==2.4.6
django-tastypie==0.10.0
greenlet==0.4.0
mechanize==0.2.5
mimeparse==0.1.3
pygeocoder==1.2.1.1
python-dateutil==2.1
requests==1.1.0
simplejson==3.3.1
six==1.4.1
wsgiref==0.1.2
